Update and rename inset-page.js to insetPage.js
